Helper view for vastaanotot
newest_vastaanotto_events view joins the vastaanotot, hakukohteet and
henkiloviitteet tables and resolves the latest vastaanotto event for
every person and persons linked to that person. Thus in this view a
vastaanotto for person A show also as a vastaanotto for person B if A
and B are linked together.