Try to use Either consistently
Functions should throw exceptions only on error situations, that is,
when some precondition has failed and the correct solution is to fix the
code. On exceptional situations an descriptive exception should be
wrapped in a Left and the return type should be Either. This indicates
to the caller that some exception handling is required.