Use 'with' query for query oiptimization when jonosijat table is used.
The psql optimizer seems to break when larger sets of hakemus (jonosijat, valinnantulokset, pistetiedot) are inserted.
This was fixed with setting autovacuums, but it doesn't work until the vacuum has completed, creating a tiny race condition
when fethcing hakemus records. Thus, jonosijat are fetched for each query first, filtering them by sijoitteluajoid and
creating a much smaller inner join for the query later.